Added some for code table entities

+++ src/main/java/com/cjs/jworks/generator/generator/impl/FileGeneratorImpl.java
@@ -5,12 +5,17 @@ import com.cjs.jworks.generator.generator.base.FileGenerator;
 
 import java.io.File;
 import java.io.FileWriter;
+import java.io.IOException;
 
 public class FileGeneratorImpl implements FileGenerator {
     @Override
     public File generate(final FileMeta fileMeta) throws Exception {
         if (fileMeta != null) {
-            final File file = new File(new File(fileMeta.getPath()), fileMeta.getName());
+            final File folder = new File(fileMeta.getPath());
+            if (!folder.exists()) {
+                throw new IOException("The system cannot find the path specified");
+            }
+            final File file = new File(fileMeta.getName());
             if (!file.exists()) {
                 file.createNewFile();
                 try (FileWriter writer = new FileWriter(file)) {
